refactoring to cleanup system metrics too
